requirements.txt
ipdb
python-dotenv
requests
zeep
flask
